requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
geographiclib==2.0
geopy==2.4.1
idna==3.6
PyYAML==6.0.1
requests==2.31.0
urllib3==2.2.1
